pyproject.toml
[project]
name = "trips-map"
version = "0.1.0"
description = " A tool to visualize your trips "
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "folium>=0.19.5",
    "geopy>=2.4.1",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
    "streamlit-folium>=0.24.0",
    "watchdog>=6.0.0",
]
